+++ tests/imports_checks/test_evalml_imports.py
@@ -1,4 +1,10 @@
 def test_evalml_imports():
-    from tabular_orchestrated.evalml import EvalMLPredict, EvalMLSearch, EvalMLSelectPipeline  # noqa: F401
+    from tabular_orchestrated.evalml import (  # noqa
+        EvalMLAnalysis,
+        EvalMLAnalysisV2,
+        EvalMLPredict,
+        EvalMLSearch,
+        EvalMLSelectPipeline,
+    )
 
     assert True
